Tested Serial Adapter code - Details of fixes in the Summary
PublishMessageFromSource - Removed if condition to check sp.IsOpen and
performing a sp.Close followed by reinitialization of sp properties -
this was repeating for every message received
and was causing error:
Exception: Exception Type: System.InvalidOperationException
Exception Message: 'PortName' cannot be set while the port is open.
Moved sp reinitialization code into new condition to check, if sp is not
intialized then setup all poperties and Open - one time only

serialport_DataReceived - Removed if NOT condition to check sp.IsOpen
and performing a sp.Open
and was causing error:
Exception: Exception Type: System.InvalidOperationException
Exception Message: The port is closed. or Access to the port 'COM2' is
denied.
Additionally, the adapter was publising a lot of null or empty messages
after the actual bytes of data were sent, causing a lot of empty
messages in the bus.
Added condition to check if message is null or empty and then Publish.